+++ src/models/testimonial.models.ts
@@ -1,8 +1,0 @@
-import { model, Schema } from "mongoose";
-
-export interface ITestimonial {
-  _id?: string;
-  userId: string;
-  testimonial: string;
-  rating: 0 | 1 | 2 | 3 | 4 | 5;
-}